COMMERCE-7229 Refactor CPMedia Implementations to package com.liferay.commerce.product.content.util
