+++ modules/apps/commerce/commerce-frontend-taglib/src/main/resources/META-INF/resources/gallery/init.jsp
@@ -17,7 +17,7 @@
 <%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
-<%@ page import="com.liferay.commerce.product.catalog.CPMedia" %><%@
+<%@ page import="com.liferay.commerce.product.content.util.CPMedia" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.json.JSONFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.json.JSONSerializer" %><%@
