+++ modules/apps/commerce/commerce-product-content-api/src/main/java/com/liferay/commerce/product/content/util/CPContentHelper.java
@@ -17,7 +17,6 @@ package com.liferay.commerce.product.content.util;
 import aQute.bnd.annotation.ProviderType;
 
 import com.liferay.commerce.product.catalog.CPCatalogEntry;
-import com.liferay.commerce.product.catalog.CPMedia;
 import com.liferay.commerce.product.catalog.CPSku;
 import com.liferay.commerce.product.content.render.CPContentRenderer;
 import com.liferay.commerce.product.model.CPDefinitionSpecificationOptionValue;
