+++ modules/apps/commerce/commerce-product-content-web/src/main/resources/META-INF/resources/init.jsp
@@ -37,7 +37,6 @@ page import="com.liferay.commerce.account.model.CommerceAccount" %><%@
 page import="com.liferay.commerce.constants.CommerceWebKeys" %><%@
 page import="com.liferay.commerce.context.CommerceContext" %><%@
 page import="com.liferay.commerce.product.catalog.CPCatalogEntry" %><%@
-page import="com.liferay.commerce.product.catalog.CPMedia" %><%@
 page import="com.liferay.commerce.product.catalog.CPSku" %><%@
 page import="com.liferay.commerce.product.constants.CPWebKeys" %><%@
 page import="com.liferay.commerce.product.content.constants.CPContentWebKeys" %><%@
@@ -46,6 +45,7 @@ page import="com.liferay.commerce.product.content.render.list.CPContentListRende
 page import="com.liferay.commerce.product.content.render.list.entry.CPContentListEntryRenderer" %><%@
 page import="com.liferay.commerce.product.content.util.CPCompareContentHelper" %><%@
 page import="com.liferay.commerce.product.content.util.CPContentHelper" %><%@
+page import="com.liferay.commerce.product.content.util.CPMedia" %><%@
 page import="com.liferay.commerce.product.content.web.internal.constants.CPCompareContentConstants" %><%@
 page import="com.liferay.commerce.product.content.web.internal.constants.CPCompareContentMiniConstants" %><%@
 page import="com.liferay.commerce.product.content.web.internal.constants.CPContentPortletConstants" %><%@
